# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZuSu Trading API",
    "version": "0.1.0",
    "description": "Day-trading automation platform. Phase 1 covers identity, portfolios, the demo broker, the trading gate and the audit log. Endpoints that would place orders do not exist yet — they arrive with the risk engine and order manager."
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zusu_at"
      }
    },
...
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Sign in with email and password",
        "tags": ["auth"],
        "description": "Returns a session, or an MFA challenge when the account requires a second factor. Administrators must complete MFA enrolment before a session is issued.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "totp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": ["email", "password"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": ["MFA_REQUIRED"]
                        },
                        "mfaToken": {
                          "type": "string"
                        }
                      },
                      "required": ["status", "mfaToken"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": ["MFA_ENROLMENT_REQUIRED"]
                        },
                        "mfaToken": {
                          "type": "string"
                        }
                      },
                      "required": ["status", "mfaToken"],
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": ["AUTHENTICATED"]
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "email": {
                              "type": "string",
                              "format": "email"
                            },
                            "displayName": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": ["ADMIN", "MANAGER", "CLIENT", "VIEWER"]
                            },
                            "mfaEnabled": {
                              "type": "boolean"
                            },
                            "permissions": {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            "portfolioIds": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "format": "uuid"
                              }
                            }
                          },
                          "required": [
                            "id",
                            "email",
                            "displayName",
                            "role",
                            "mfaEnabled",
                            "permissions",
                            "portfolioIds"
                          ],
                          "additionalProperties": false
                        },
                        "csrfToken": {
                          "type": "string"
                        },
                        "accessTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": ["status", "user", "csrfToken", "accessTokenExpiresAt"],
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/mfa/enrol": {
      "post": {
        "summary": "Start multi-factor enrolment using a login challenge token",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mfaToken": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": ["mfaToken"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secret": {
                      "type": "string"
                    },
                    "otpauthUrl": {
                      "type": "string"
                    },
                    "qrDataUrl": {
                      "type": "string"
                    }
                  },
                  "required": ["secret", "otpauthUrl", "qrDataUrl"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/mfa/verify": {
      "post": {
        "summary": "Complete an MFA challenge and receive a session",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mfaToken": {
                    "type": "string",
                    "minLength": 1
                  },
                  "totp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": ["mfaToken", "totp"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["AUTHENTICATED"]
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": ["ADMIN", "MANAGER", "CLIENT", "VIEWER"]
                        },
                        "mfaEnabled": {
                          "type": "boolean"
                        },
                        "permissions": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "portfolioIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          }
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "role",
                        "mfaEnabled",
                        "permissions",
                        "portfolioIds"
                      ],
                      "additionalProperties": false
                    },
                    "csrfToken": {
                      "type": "string"
                    },
                    "accessTokenExpiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": ["status", "user", "csrfToken", "accessTokenExpiresAt"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/mfa/activate": {
      "post": {
        "summary": "Confirm an authenticator code and switch MFA on",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "totp": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": ["totp"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mfaEnabled": {
                      "type": "boolean",
                      "enum": [true]
                    }
                  },
                  "required": ["mfaEnabled"],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "this deployment"
    }
  ],
...
}